🔨 💥 Make watch()/unwatch() async (#36)
On the big, complex folder structures, the watch() and potentially unwatch() start to take a lot to time blocking the asyncio loop.

In this PR, we make them async to account for such cases.

Additionally, I had to upgrade a bunch of GH Actions & Python dependencies to make pdm install work.